Move common libraries to internal
