+++ cmd/b1gtool/update-games.go
@@ -9,8 +9,8 @@ import (
 	"strconv"
 
 	fs "cloud.google.com/go/firestore"
-	"github.com/reallyasi9/b1gpickem/cfbdata"
-	"github.com/reallyasi9/b1gpickem/firestore"
+	"github.com/reallyasi9/b1gpickem/internal/cfbdata"
+	"github.com/reallyasi9/b1gpickem/internal/firestore"
 )
 
 // ugFlagSet is a flag.FlagSet for parsing the update-games subcommand.
