+++ cmd/b1gtool/update-sagarin.go
@@ -17,7 +17,7 @@ import (
 	"time"
 
 	fs "cloud.google.com/go/firestore"
-	"github.com/reallyasi9/b1gpickem/firestore"
+	"github.com/reallyasi9/b1gpickem/internal/firestore"
 )
 
 const ratingColor = "#9900ff"
