+++ internal/cfbdata/collections.go
