+++ internal/cfbdata/game.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	fs "cloud.google.com/go/firestore"
-	"github.com/reallyasi9/b1gpickem/firestore"
+	"github.com/reallyasi9/b1gpickem/internal/firestore"
 )
 
 type Game struct {
