+++ internal/cfbdata/venue.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 
 	fs "cloud.google.com/go/firestore"
-	"github.com/reallyasi9/b1gpickem/firestore"
+	"github.com/reallyasi9/b1gpickem/internal/firestore"
 )
 
 type Venue struct {
