+++ internal/firestore/slate.go
