+++ internal/firestore/streak.go
