+++ internal/firestore/week.go
